requirements.txt
blinker==1.9.0
click==8.3.2
colorama==0.4.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.8
supabase==2.3.4
python-dotenv==1.0.1
pytest==8.0.0
pytest-flask==1.3.0
